requirements.txt
attrs==20.3.0
boto3==1.17.14
botocore==1.20.14
certifi==2020.12.5
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
Fiona==1.8.18
geopandas==0.9.0
jmespath==0.10.0
logging==0.4.9.6
munch==2.5.0
numpy==1.20.1
pandas==1.2.3
psycopg2==2.8.6
pyproj==3.0.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
s3transfer==0.3.4
Shapely==1.7.1
six==1.15.0
urllib3==1.26.3
